poll_backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
mysql-connector-python==8.3.0
PyJWT==2.8.0
Werkzeug==2.3.7
pytest==7.4.0
pytest-env==0.8.2
pytest-flask==1.2.0
coverage==7.3.1
flake8==6.1.0 